AplikacjaTestowanie/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
requests 
pydantic
pytest
pytest-cov
httpx
python-dotenv

